refactor: Extract app config mergers
For the merge util to be used outside cozy-scripts (Banks)
